id array

+++ src/main/java/cn/nukkit/level/format/anvil/Chunk.java
@@ -12,7 +12,6 @@ import cn.nukkit.level.format.generic.EmptyChunkSection;
 import cn.nukkit.nbt.NBTIO;
 import cn.nukkit.nbt.tag.*;
 import cn.nukkit.utils.*;
-
 import java.io.ByteArrayInputStream;
 import java.io.DataInputStream;
 import java.io.IOException;
@@ -276,7 +275,7 @@ public class Chunk extends BaseChunk {
             }
             CompoundTag s = new CompoundTag(null);
             s.putByte("Y", section.getY());
-            s.putByteArray("Blocks", section.getIdArray());
+            s.putIntArray("Blocks", section.getIdArray());
             s.putByteArray("Data", section.getDataArray());
             s.putByteArray("BlockLight", section.getLightArray());
             s.putByteArray("SkyLight", section.getSkyLightArray());
@@ -359,7 +358,7 @@ public class Chunk extends BaseChunk {
             }
             CompoundTag s = new CompoundTag(null);
             s.putByte("Y", (section.getY()));
-            s.putByteArray("Blocks", section.getIdArray());
+            s.putIntArray("Blocks", section.getIdArray());
             s.putByteArray("Data", section.getDataArray());
             s.putByteArray("BlockLight", section.getLightArray());
             s.putByteArray("SkyLight", section.getSkyLightArray());
